move out static guided tour and make immutable
